refactor : /type  --> /data/type
type 패키지를 data 패키지 하위로 이동

jira issue key : SB-9
