+++ board/src/main/java/kim/zhyun/board/config/SwaggerConfig.java
@@ -1,0 +1,4 @@
+package kim.zhyun.board.config;
+
+public class SwaggerConfig {
+}
